refactor: highlightFirstLineInTextView 메서드를 UITextView의 익스텐션으로 이동
